pyproject.toml
[tool.poetry]
name = "nep-shop"
version = "0.1.0"
description = ""
authors = ["sudomanoj <manoj.765421@bumc.tu.edu.np>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = {extras = ["blacklist"], version = "^5.4.0"}
django-jazzmin = "^3.0.1"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
